refactor: axis optimization
