+++ packages/vtable/src/layout/chart-helper/get-chart-spec.ts
@@ -9,9 +9,9 @@ import { Factory } from '../../core/factory';
 import type { GetAxisDomainRangeAndLabels } from './get-axis-domain';
 import { DEFAULT_TEXT_FONT_SIZE } from '../../components/axis/get-axis-attributes';
 import { convertDomainToTickData } from '@src/vrender';
-import { getTickModeFunction, getZeroAlignTickAlignTicks } from './tick-align';
+import { generateChartAxesConfig, NO_AXISID_FRO_VTABLE } from './axis-utils';
 
-const NO_AXISID_FRO_VTABLE = 'NO_AXISID_FRO_VTABLE';
+// NO_AXISID_FRO_VTABLE 常量已移至 types.ts
 
 export function getRawChartSpec(col: number, row: number, layout: PivotHeaderLayoutMap): any {
   const paths = layout.getCellHeaderPaths(col, row);
@@ -184,190 +184,8 @@ export function getChartSpec(col: number, row: number, layout: PivotHeaderLayout
 }
 
 export function getChartAxes(col: number, row: number, layout: PivotHeaderLayoutMap): any {
-  const axes: any[] = [];
-  if (layout.indicatorsAsCol) {
-    // const indicatorKeys = layout.getIndicatorKeyInChartSpec(col, row).slice(0, 2);
-    const indicatorKeys = layout.getIndicatorKeyInChartSpec(col, row);
-    // const colIndex = layout.getRecordIndexByCol(col);
-    const colPath = layout.getColKeysPath(col, row);
-    indicatorKeys.forEach((key, index) => {
-      const { range, targetTicks, targetRange, axisOption } = getAxisRangeAndTicks(
-        col,
-        row,
-        index,
-        index === 0 ? 'bottom' : 'top',
-        index === 0 ? 'top' : 'bottom',
-        indicatorKeys,
-        colPath,
-        layout
-      );
-      if (isNumber(axisOption?.min)) {
-        (range as any).min = axisOption.min;
-      }
-      if (isNumber(axisOption?.max)) {
-        (range as any).max = axisOption.max;
-      }
-
-      if (hasSameAxis(axisOption, axes)) {
-        return;
-      }
-
-      axes.push(
-        merge(
-          {
-            range,
-            label: { style: { fontSize: DEFAULT_TEXT_FONT_SIZE } }
-          },
-          axisOption,
-          {
-            type: axisOption?.type || 'linear',
-            orient: index === 0 ? 'bottom' : 'top',
-            // visible: true,
-            label: { visible: false, flush: true },
-            // label: { flush: true },
-            title: { visible: false },
-            domainLine: { visible: false },
-            seriesIndex: axisOption?.seriesId ? undefined : index,
-            // height: -1,
-            tick: {
-              tickMode: getTickModeFunction(targetTicks, targetRange, range, index)
-            },
-            sync: { axisId: NO_AXISID_FRO_VTABLE } // hack for fs
-          }
-        )
-      );
-    });
-
-    let rowDimensionKey = layout.getDimensionKeyInChartSpec(layout.rowHeaderLevelCount, row);
-    if (isArray(rowDimensionKey)) {
-      rowDimensionKey = rowDimensionKey[0];
-    }
-    const data =
-      layout.dataset.cacheCollectedValues[rowDimensionKey] ||
-      layout.dataset.collectedValues[rowDimensionKey] ||
-      ([] as string[]);
-    const rowPath = layout.getRowKeysPath(col, row);
-    const domain = (data as any)[rowPath ?? ''] as Set<string>;
-    const { axisOption, isPercent, chartType } = getAxisOption(col, row, 'left', layout);
-    axes.push(
-      // 左侧维度轴
-      merge(
-        {
-          domain: axisOption?.type === 'linear' && !Array.isArray(domain) ? undefined : Array.from(domain ?? []),
-          range: axisOption?.type === 'linear' && !Array.isArray(domain) ? domain : undefined,
-          label: { style: { fontSize: DEFAULT_TEXT_FONT_SIZE } }
-        },
-        axisOption,
-        {
-          type: axisOption?.type ?? 'band',
-          orient: 'left',
-          label: { visible: false },
-          domainLine: { visible: false },
-          tick: { visible: false },
-          subTick: { visible: false },
-          title: { visible: false }
-        }
-      )
-    );
-  } else {
-    // const indicatorKeys = layout.getIndicatorKeyInChartSpec(col, row).slice(0, 2);
-    const indicatorKeys = layout.getIndicatorKeyInChartSpec(col, row);
-    const rowPath = layout.getRowKeysPath(col, row);
-    indicatorKeys.forEach((key, index) => {
-      const { range, targetTicks, targetRange, axisOption } = getAxisRangeAndTicks(
-        col,
-        row,
-        index,
-        index === 0 ? 'left' : 'right',
-        index === 0 ? 'right' : 'left',
-        indicatorKeys,
-        rowPath,
-        layout
-      );
-      if (isNumber(axisOption?.min)) {
-        (range as any).min = axisOption.min;
-      }
-      if (isNumber(axisOption?.max)) {
-        (range as any).max = axisOption.max;
-      }
-
-      if (hasSameAxis(axisOption, axes)) {
-        return;
-      }
-      const { chartType } = getAxisOption(col, row, index === 0 ? 'left' : 'right', layout);
-      let domain: Array<string> = [];
-      if (chartType === 'heatmap') {
-        //为heatmap时 需要获取维度轴的domain 因为有可能都是离散轴。这里的处理对应get-axis-config.ts中的getAxisConfigInPivotChart方法处理
-        const rowDimensionKey = layout.getDimensionKeyInChartSpec(layout.rowHeaderLevelCount, row, 'yField');
-        const data = layout.dataset.collectedValues[rowDimensionKey] ?? ([] as string[]);
-
-        const rowPath = layout.getRowKeysPath(col, row);
-        domain = ((data as any)?.[rowPath ?? ''] as Array<string>) ?? [];
-      }
-      axes.push(
-        merge(
-          {
-            range,
-            label: { style: { fontSize: DEFAULT_TEXT_FONT_SIZE } },
-            domain: axisOption?.type === 'linear' ? undefined : Array.from(domain)
-          },
-          axisOption,
-          {
-            type: axisOption?.type || 'linear',
-            orient: index === 0 ? 'left' : 'right',
-            // visible: true,
-            label: { visible: false, flush: true },
-            // label: { flush: true },
-            title: { visible: false },
-            domainLine: { visible: false },
-            seriesIndex: axisOption?.seriesId ? undefined : index,
-            // width: -1,
-            // grid: index === 0 ? undefined : { visible: false }
-            tick: {
-              tickMode: getTickModeFunction(targetTicks, targetRange, range, index),
-              visible: false // 轴刻度不显示
-            },
-            sync: { axisId: NO_AXISID_FRO_VTABLE } // hack for fs
-          }
-        )
-      );
-    });
-
-    let columnDimensionKey = layout.getDimensionKeyInChartSpec(col, layout.columnHeaderLevelCount);
-    if (isArray(columnDimensionKey)) {
-      columnDimensionKey = columnDimensionKey[0];
-    }
-    const data =
-      layout.dataset.cacheCollectedValues[columnDimensionKey] ||
-      layout.dataset.collectedValues[columnDimensionKey] ||
-      ([] as string[]);
-    const colPath = layout.getColKeysPath(col, row);
-    const domain: string[] | Set<string> = ((data as any)?.[colPath ?? ''] as Set<string>) ?? [];
-
-    const { axisOption, isPercent, chartType } = getAxisOption(col, row, 'bottom', layout);
-    axes.push(
-      // 底部维度轴
-      merge(
-        {
-          domain: axisOption?.type === 'linear' && !Array.isArray(domain) ? undefined : Array.from(domain ?? []),
-          range: axisOption?.type === 'linear' && !Array.isArray(domain) ? domain : undefined,
-          label: { style: { fontSize: DEFAULT_TEXT_FONT_SIZE } }
-        },
-        axisOption,
-        {
-          type: axisOption?.type ?? 'band',
-          orient: 'bottom',
-          visible: true,
-          label: { visible: false },
-          domainLine: { visible: false },
-          tick: { visible: false },
-          subTick: { visible: false },
-          title: { visible: false }
-        }
-      )
-    );
-  }
-  return axes;
+  // 使用共享函数生成坐标轴配置
+  return generateChartAxesConfig(col, row, layout);
 }
 /**
  *  获取单元格对应spec的dataId。
@@ -420,26 +238,4 @@ export function checkHasChart(layout: PivotHeaderLayoutMap | SimpleHeaderLayoutM
   return isHasChart;
 }
 
-function hasSameAxis(axisOption: any, axes: any[]) {
-  if (axisOption && isArray(axisOption.seriesId) && axisOption.seriesId.length > 0) {
-    // find same seriesId axes
-    const sameSeriesIdAxes = (axes as any[]).filter(axis => {
-      // same seriesId
-      if (
-        axis.orient === axisOption.orient &&
-        axis.seriesId &&
-        axis.seriesId.length === axisOption.seriesId.length &&
-        axis.seriesId.every((id: string, index: number) => id === axisOption.seriesId[index])
-      ) {
-        return true;
-      }
-      return false;
-    });
-
-    if (sameSeriesIdAxes.length > 0) {
-      // has same seriesId axes
-      return true;
-    }
-  }
-  return false;
-}
+// hasSameAxis 函数已移至 axis-utils.ts
